Create a Postgres-specific implementation of DBALEventStore
Due to PostgreSQL's default of setting lowercase on column names, the
standard implementation of DBALEventStore would not work with
Postgres (See Issue #52).

Fixing the quoting to work with PosgreSQL would then break the class
such that it would no longer work wit MySQL. Thus, a new version was
needed that would work with Postgres, allowing the original to
continue working as it was.

This commit changes a few properties and methods of DBALEventStore
from private to protected, and adds a PostgresDBALEventStore class
which extends DBALEventStore and overrides the methods where the
column quoting is needed. No other implementation changes were made
to the class.