Fixes #52 - Tweak to SELECT statement in DBALEventStore for PostgreSQL
PostgreSQL forces all column names to lowercase. This causes issues with
the recordedOn colum in the SELECT statement, as the row map winds up with
parameters that are all lowercase instead of the required camel case for
the recordedOn column, and then causes errors elsewhere in the class.

This fix forces the selected column be returned with the camel case
name. This was the simplest solution that still allows the class to
work with PostgreSQL, MySQL, and Sqlite.

Other solutions would have required complex quoting schemes, which would
work for some database packages but break others. So short of creating
a subclass for PostgreSQL, this was the simple way to go.