Tweak to SELECT statement in DBALEventStore for PostgreSQL - Fixes #52
PostgreSQL forces all column names to lowercase. This causes issues with
the recordedOn colum in the SELECT statement, as the row map winds up with
parameters that are all lowercase instead of the required camel case for
the recordedOn column, and then causes errors elsewhere in the class.

This fix changes the column name the be recorded_on instead of recordedOn,
so that it is no longer an issue. This change is for the database only and does
not affect the key name in any domain object.