Extract reusable ElasticSearch client handling from the implementation
This could prove to be useful in the case that someone wants to leverage
existing ElasticSearch code without extending the base ElasticSearch repository
implementations. In particular, this is useful if ssomeone does not have a need
or desire to extend the base class but still wants to use some of its tools
without having to copy & paste a LOT of code.

refs #109