Make the simple command bus more resilient.
If an exception occurs during the handling of a command, the simple
command bus becomes unusable (`$isDispatching` stays `true`).
With this change the bus will be able to dispatch subsequent commands
even if an exception occurs.

Thanks for @kimlai for the work on this.
I just added the test case.

This also closes https://github.com/qandidate-labs/broadway/pull/82