[bundle] Make dbal event store optional
If a project does not use the dbal event store the bundle should not
require its configuration and load its configuration anyway. This commit
adds a new configuration option to disable the dbal event store
completely. In that case the `broadway.event_store.dbal` service will
not be available and the `broadway.event_store` alias will not be set
either, so developers can set the alias in their own application.