Provide access to Metadata values with `get()` and `all()`
Before this commit it wasn't possible to read values from Metadata
except when you serialized it.

Eventhough `serialize()` is functionally equal to `all()` it was
decided at #318 to add it for semantics sake. In other words,
`serialize()` must be used for persisting the Metadata, `all()`
or `get()` must be used to act upon.